give a try to more complex queue in dijkstra algorithm
Tried to add the walk connection to the queue in a lazy way, so
connection that are very long would not be touched even once.
Not sure if the lazy_queue is working properly, but for now results
are worse than before.

+++ Components/Public_Transport/Preprocessing_data/data_reader.cpp
@@ -28,7 +28,7 @@ void read_stops(string &path, Data *pt_data){
         longtitude_max = max(longtitude_max, (int)parts[4].size() + 1);
         pt_data->stops.push_back(Stop(new_id, parts[2], parts[3], parts[4]));
     }
-    stops_lim = pt_data->stops.size();
+    stops_lim = (int)pt_data->stops.size();
     for(int i = 0; i < stops_lim; i++){
         pt_data->stops[i].connections.resize(stops_lim);
     }
@@ -54,7 +54,7 @@ void read_trips(string &path, Data *pt_data, int day_id){
 
         pt_data->trips.push_back(Trip(new_id, parts[0]));
     }
-    trips_lim = pt_data->trips.size();
+    trips_lim = (int)pt_data->trips.size();
 }
 
 void read_stop_times(string &path, Data *pt_data, int day_id){
